# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Laravel with Passport Auth Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "Forename",
                                    "Surname",
                                    "Email",
                                    "Password",
                                    "Category",
                                    "ConfirmPassword",
                                    "Username"
                                ],
                                "properties": {
                                    "Forename": {
                                        "type": "string"
                                    },
                                    "Surname": {
                                        "type": "string"
                                    },
                                    "DisplayName": {
                                        "type": "string"
                                    },
                                    "Username": {
                                        "type": "string"
                                    },
                                    "Email": {
                                        "type": "string"
                                    },
                                    "Password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "ConfirmPassword": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "Category": {
                                        "type": "string",
                                        "enum": [
                                            "-",
                                            "Male",
                                            "Female",
                                            "Trans"
                                        ]
                                    },
                                    "PhoneNumber": {
                                        "type": "string"
                                    },
                                    "TwoFactor": {
                                        "type": "string",
                                        "default": "No",
                                        "enum": [
                                            "Yes",
                                            "No"
                                        ]
                                    },
                                    "AgreeTerms": {
                                        "type": "string",
                                        "default": "Yes",
                                        "enum": [
                                            "Yes",
                                            "No"
                                        ]
                                    },
                                    "YearsOld": {
                                        "type": "string",
                                        "default": "Yes",
                                        "enum": [
                                            "Yes",
                                            "No"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "remember_me": {
                                        "type": "string",
                                        "default": "Yes",
                                        "enum": [
                                            "Yes",
                                            "No"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/oauth/token/refresh",
                        "scopes": {}
                    }
                },
                "scheme": "https"
            }
        }
    },
    "security": [
        []
    ]
}